old/pyproject.toml
[project]
name = "multirig"
version = "0.1.0"
description = "Control and sync two ham radio rigs via hamlib with a modern web UI"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "pyserial>=3.5",
    "orjson>=3.9.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "rjsmin>=1.2.0",
    "rcssmin>=1.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["multirig*"]
exclude = ["ext*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["ext", ".venv", "build", "dist"]
addopts = ["--ignore=ext"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-playwright>=0.5.0",
    "psutil>=5.9.0",
    "httpx>=0.27.0",
]

[project.scripts]
multirig = "multirig.app:run"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
